Add 09-02-2015 MeetUp information & change 'hidemap' to 'online'.
This change adds the 09-02-2015 MeetUp information. But, furthermore it
changes the 'hidemap' toggle to an 'online' toggle. This allows us to
toggle the map off and the "Ask us a Question" and "Watch Live" buttons
on whenever 'online' = true. It shows the map (and adds the google map
script) and hides the "Ask" and "Watch" buttons since they don't really
pertain to an in-person meetup.