May 2016 meetup at lupes (#16)
* Update location address information.

* Update date and talk information.

* Update speaker information.

Additionally, this commit allows for an empty speaker in which case the default LaravelDFW image and "Presented by LaravelDFW" language are used.

* Remove additional information message.

Additionally, this commit allows for empty additional information so that you know that is an option. If empty, however, the message will not be displayed. Otherwise, it will.

* Update lat/long information.

Additionally, this commit moves the lat/long assignment into the routes file so it lives in proximity to all the other location information.

* Update the RSVP url to point at the next MeetUp.